move around vltava files
