+++ examples/components/vltava/src/components/vltava/index.ts
@@ -3,5 +3,4 @@
  * Licensed under the MIT License.
  */
 
-export * from "./tabs";
 export * from "./vltava";
